task completed

+++ src/products/products.service.ts
@@ -4,20 +4,22 @@ import { Repository } from 'typeorm';
 import { Product } from '../entities/products.entity';
 import { CreateProductDto } from './dto/product.dto';
 import { UpdateProductDto } from './dto/product.dto';
+import * as fs from "node:fs";
 
 @Injectable()
 export class ProductsService {
     constructor(
         @InjectRepository(Product)
         private readonly productRepository: Repository<Product>,
-    ) {}
+    ) {
+    }
 
     async findAll(): Promise<Product[]> {
         return this.productRepository.find();
     }
 
     async findOne(id: number): Promise<Product> {
-        const product = await this.productRepository.findOne({ where: { id } });
+        const product = await this.productRepository.findOne({where: {id}});
         if (!product) {
             throw new NotFoundException(`Product with ID ${id} not found`);
         }
@@ -42,7 +44,14 @@ export class ProductsService {
     }
 
     async remove(id: number): Promise<void> {
-        const product = await this.findOne(id);
+        const product = await this.productRepository.findOne({where: {id: id}});
+        if (!product) {
+            throw new NotFoundException(`Product with ID ${id} not found`);
+        }
+        if (product.image) {
+            fs.unlinkSync(product.image);
+        }
+
         await this.productRepository.remove(product);
     }
 }
